init commit with loads of code
First commit for the first day of Advent of Code!
Time to practice some java, because it's fun.

I'm trying to do a simple TDD approach where I write a test
module/method for my other methods. Here we have already
made a method to calculate the fuel needed for Santas spaceship
and we are going to test it for the 4 given masses in the
exercise/assignment.