remove TestMasses and add TotalFuelCalculations
We have removed the method TestMasses since it just compared
the test answer with the correct answer, so it got replaced with
a simple == operation.

We have started a new method for calculating the correct fuel
needed based on the assignment.

We could do something recursive with this, but it could
also be solved with a while-loop.

Or both.